requirements/prod.txt
fastapi
uvicorn[standard]
gunicorn
sqlmodel
pydantic
alembic
pyodbc
azure-storage-blob
azure-identity
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pytz
sentry-sdk